+++ src/imports/copypaste/index.ts
@@ -1,0 +1,7 @@
+import CopyPasteProvider from './CopyPasteProvider';
+
+export default {
+	__init__: ['copyPasteProvider'],
+	copyPasteProvider: ['type', CopyPasteProvider],
+};
+